[feat] add new strategy & refactor
Added the `/watch` endpoint strategy. Descriptions and related
videos now work without a proxy server.

Also did some refactoring, to make it more straight forward adding more endpoints.
