+++ src/components/endpoints/index.js
@@ -1,0 +1,3 @@
+export { default as innertube } from './innertube.js';
+export { default as proxy } from './proxy.js';
+export { default as watch } from './watch.js';
