+++ src/utils/index.js
@@ -1,4 +1,5 @@
 import { nativeJSONParse } from './natives';
+import { generateSha1Hash } from './sha1';
 
 export const isDesktop = window.location.host !== 'm.youtube.com';
 export const isMusic = window.location.host === 'music.youtube.com';
@@ -63,6 +64,51 @@ export function pageLoaded() {
 
 export function createDeepCopy(obj) {
     return nativeJSONParse(JSON.stringify(obj));
+}
+
+export function getYtcfgValue(name) {
+    return window.ytcfg?.get(name);
+}
+
+export function getSignatureTimestamp() {
+    return (
+        getYtcfgValue('STS') ||
+        (() => {
+            // STS is missing on embedded player. Retrieve from player base script as fallback...
+            const playerBaseJsPath = document.querySelector('script[src*="/base.js"]')?.src;
+
+            if (!playerBaseJsPath) return;
+
+            const xmlhttp = new XMLHttpRequest();
+            xmlhttp.open('GET', playerBaseJsPath, false);
+            xmlhttp.send(null);
+
+            return parseInt(xmlhttp.responseText.match(/signatureTimestamp:([0-9]*)/)[1]);
+        })()
+    );
+}
+
+export function getSidCookie() {
+    return getCookie('SAPISID') || getCookie('__Secure-3PAPISID');
+}
+
+export function generateSidBasedAuth() {
+    const sid = getSidCookie();
+    const timestamp = Math.floor(new Date().getTime() / 1000);
+    const input = timestamp + ' ' + sid + ' ' + location.origin;
+    const hash = generateSha1Hash(input);
+    return `SAPISIDHASH ${timestamp}_${hash}`;
+}
+
+export function isUserLoggedIn() {
+    // Session Cookie exists?
+    if (!getSidCookie()) return false;
+
+    // LOGGED_IN doesn't exist on embedded page, use DELEGATED_SESSION_ID as fallback
+    if (typeof getYtcfgValue('LOGGED_IN') === 'boolean') return getYtcfgValue('LOGGED_IN');
+    if (typeof getYtcfgValue('DELEGATED_SESSION_ID') === 'string') return true;
+
+    return false;
 }
 
 export function getCurrentVideoStartTime(currentVideoId) {
